refactored sseInitialization into its own utils file and imported into reduxcomparison to improve readability
